backend/requirements.txt
fastapi>=0.100.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
alembic>=1.11.0
uvicorn>=0.23.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
bcrypt>=4.0.1
asyncpg>=0.28.0
loguru>=0.7.0
email-validator>=1.3.1 